R6 Class Representing a Fish

A fish object has all the life history data needed to model a finfish population

Details

Fish like to swim

Methods

Public methods

Fish$new()

Fish$plot()

Fish$swim()

Fish$clone()

Method new()

fill in fish object

Usage

Fish$new(
  common_name = NA,
  scientific_name = NA,
  linf = NA,
  vbk = NA,
  t0 = -0.1,
  cv_len = 0.1,
  length_units = "cm",
  min_age = 0,
  max_age = NA,
  weight_a = NA,
  weight_b = NA,
  pups = 10,
  weight_units = "kg",
  fec_form = "weight",
  fec_expo = 1,
  length_50_mature = NA,
  length_95_mature = NA,
  delta_mature = 0.1,
  age_50_mature = NA,
  age_95_mature = NA,
  age_mature = NA,
  length_mature = NA,
  m = NA,
  steepness = 0.8,
  r0 = 10000,
  ssb0 = NA,
  density_dependence_form = 1,
  adult_movement = 0,
  adult_movement_sigma = 4,
  recruit_movement = 0,
  recruit_movement_sigma = 10,
  query_fishlife = T,
  sigma_r = 0,
  rec_ac = 0,
  cores = 4,
  mat_mode = "age",
  default_wb = 2.8,
  tune_weight = FALSE,
  density_movement_modifier = 1,
  linf_buffer = 1.2,
  resolution = NA,
  seasonal_habitat = list(),
  season_blocks = list(),
  recruit_habitat = NA,
  fished_depletion = 1,
  rec_form = 1,
  burn_years = 50,
  seasonal_hab = NA,
  seasons = 1,
  explt_type = "f",
  init_explt = 0.1,
  get_common_name = FALSE
)

Arguments

common_name

common name

pups

fec_form

fec_expo

exponent for fecundity at weight relationship, 1 = isometric > 1 hyperallometric

recruit_movement_sigma

explt_type

SwimSwim advances the population one time step

get_common_name

Method plot()

Usage

Fish$plot(type = 2)

Method swim()

Usage

Fish$swim(
  burn_steps = 0,
  season = 1,
  f_p_a = NULL,
  last_n_p_a = NULL,
  tune_unfished = 0
)

Arguments

burn_steps

number of steps for burn in period if applicable

season

the current season

f_p_a

matrix of fishing mortality by patch and age

last_n_p_a

matrix of initial numbers by patch and age

tune_unfished

boolean indicating whether to tune unfished

Returns

the population in the next time step

Method clone()

The objects of this class are cloneable with this method.

Usage

Fish$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.