Create Critter

Creates a critter object. If only a scientific name is provided create_critter will try and look up relevant life history from FishLife

Usage

create_critter(
  common_name = NA,
  scientific_name = NA,
  get_common_name = FALSE,
  critter_type = "fish",
  habitat = list(),
  season_blocks = list(),
  recruit_habitat = NA,
  seasons = 1,
  lorenzen_c = -1,
  fec_form = "weight",
  adult_home_range = 1,
  recruit_home_range = 2,
  adult_diffusion = NULL,
  recruit_diffusion = NULL,
  depletion = 0.4,
  fished_depletion = NULL,
  init_explt = 0.1,
  f = NULL,
  explt_type = "f",
  burn_years = 50,
  weight_a = NA,
  fec_expo = 1,
  resolution = c(10, 10),
  patch_area = 1,
  spawning_seasons = NA,
  density_dependence = "global_habitat",
  growth_model = "von_bertalanffy",
  ...
)

Arguments

common_name

the common name of the species

scientific_name

the scientific name of the species, preferable to common name

get_common_name

TRUE or FALSE to lookup common name from scientific name. Requires internet connection

critter_type

placeholder for someday if non-Fish objects are implemented

habitat

a list with adult habitat per season

season_blocks

a list with seasons per block

recruit_habitat

habitat for recruitment

seasons

number of seasons per year (integer)

lorenzen_c

rate of the lorenzen curve, defaults to -1

fec_form

one of "weight" or "pups"

adult_diffusion

adult diffusion rate

recruit_diffusion

recruit diffusion rate

depletion

depletion (biomass / unfished biomass) at start of simulation

init_explt

initial annual exploitation rate (fraction of exploitable population killed)

explt_type

f or fmsy

burn_years

number of years to burn in the simulation prior to starting things

weight_a

alpha parameter in the allometric weight function alpha x length ^ beta

fec_expo

exponent for fecundity relationship. >1 means hyperallometric fecundity

resolution

the resolution of the system, either an integer or a vector integers of length two specifying the dimensions of the system in width and height (e.g. c(10,100))

patch_area

the area of each patch

spawning_seasons

which seasons spawning occurs

density_dependence

one of 'global_habitat','local_habitat','pre_dispersal','post_dispersal','global_ssb'

growth_model

one of "von_bertalanffy", "power", "growth_cessation"

...

additional parameters passed to Fish class, see ?Fish

Details

Critical inputs are adult_movement, adult_movement_sigma, and resolution

See ?Fish for documentation of all possible parameters