Optimize MPA network

Optimize MPA network

Usage

optimize_mpa(
  fauna,
  fleets,
  starting_conditions = NA,
  alpha = 0.33,
  max_prop_mpa = 1,
  resolution,
  prop_sampled = 0.2,
  max_delta = 2,
  workers = 6,
  bio_objective = "max_ssb",
  econ_objective = "yield",
  work_backwards = TRUE,
  patches_at_a_time = 1
)

Arguments

fauna

fauna object

fleets

fleet object

starting_conditions

starting conditions of the simmar object

alpha

the weight given to conservation relative to economics (range 0 to 1)

max_prop_mpa

max proportion of cells to place in MPA

resolution

the resolution of the simulated system

prop_sampled

the proportion of cells to sample in each SIR iteration

workers

number of workers for parallel process

objective

one of max_ssb to maximize spawning stock biomass or min_loss to prioritize not having any losses

Value

a list with results of MPA optimization