Plot Marlin generates time series, spatial, or length composition plots from marlin run

Plot Marlin generates time series, spatial, or length composition plots from marlin run

Usage

plot_marlin(
  ...,
  steps_to_plot = NA,
  plot_var = "ssb",
  plot_type = "time",
  fauna = NULL,
  drop_recruits = TRUE,
  plots = "fauna",
  max_scale = TRUE
)

Arguments

...

named or unnamed outputs of process_marlin

steps_to_plot

which steps to include in the plot

plot_var

the primary plotting variable, one of "c" (catch) "n" (numbers),"b" (biomass),"ssb" (spawning stock biomass)

plot_type

the plot type, one of "time", "length", or "space"

fauna

a fauna object corresponding to the critters in process_marlin outputs

drop_recruits

TRUE to drop recruits from length composition plots

Value

a plot

Examples

if (FALSE) { # \dontrun{

plot_marlin(
`MPA: Sharks Offshore` = proc_offshore_mpa,
`No MPA` = proc_nearshore,
`MPA: Sharks Nearshore` = proc_nearshore_mpa,
)

} # }