simmar is the wrapper function for the marlin package

when passed fauna and fleet objects, simmar will advance the population for a number of steps

Usage

simmar(
  fauna = list(),
  fleets = list(),
  manager = list(),
  habitat = list(),
  years = 100,
  steps = NA,
  starting_season = NA,
  initial_conditions = NA,
  starting_step = NA,
  keep_starting_step = TRUE,
  log_rec_devs = NULL,
  cor_rec = diag(length(fauna))
)

Arguments

fauna

a list of fauna objects

fleets

a list of fleet objects

manager

a list of management actions

habitat

a list of habitat over time

years

the number of years to run the simulation

steps

the number of steps to run, as an alternative to years

starting_season

the starting season for the simulation

initial_conditions

initial conditions for the simulation, in the form simmar()[final step]

starting_step

the step to start the simulation from, used to keep track of steps across multiple runs of simmar

keep_starting_step

should the starting step by kept (TRUE) or dropped (FALSE)

log_rec_devs

externally supplied log recruitment deviates

cor_rec

correlation matrix in recruitment deviates across species

Value

a list containing the results of the simulation