tune_fleets tunes parameters of the fleet model to achieve desired initial conditions. note that this is not exact: post-tuning values will not perfectly match inputs since for example some tuning steps depend on prior tuning step, making it difficult to tune everything at once.

tune_fleets tunes parameters of the fleet model to achieve desired initial conditions. note that this is not exact: post-tuning values will not perfectly match inputs since for example some tuning steps depend on prior tuning step, making it difficult to tune everything at once.

Usage

tune_fleets(
  fauna,
  fleets,
  years = 50,
  tune_type = "f",
  tune_costs = TRUE,
  fine_tune_costs = TRUE
)

Arguments

fauna

a fauna object

fleets

a fleet object

years

the number of years to tune over

tune_type

one of 'f' or 'depletion' to tune catchability to achieve a desired fishing mortality rate (f) or a target depletion (B/B0)

tune_costs

TRUE or FALSE to tune costs to a target cost to revenue ratio

fine_tune_costs

TRUE or FALSE

Value

tuned fleet object